+++ benchmark/pg-vector-selector-benchmark/tsconfig.json
@@ -71,7 +71,7 @@
     // "inlineSourceMap": true,                          /* Include sourcemap files inside the emitted JavaScript. */
     // "noEmit": true,                                   /* Disable emitting files from a compilation. */
     // "outFile": "./",                                  /* Specify a file that bundles all outputs into one JavaScript file. If 'declaration' is true, also designates a file that bundles all .d.ts output. */
-    "outDir": "./lib", /* Specify an output folder for all emitted files. */
+    // "outDir": "./lib", /* Specify an output folder for all emitted files. */
     // "declarationMap": true,                           /* Create sourcemaps for d.ts files. */
     // "emitDeclarationOnly": true,                      /* Only output d.ts files and not JavaScript files. */
     "sourceMap": true, /* Create source map files for emitted JavaScript files. */
@@ -105,5 +105,5 @@
       }
     ]
   },
-  "include": ["src"]
+  "include": ["src/**/*.ts"]
 }
